requirements.txt
apache-airflow-providers-google==10.13.0
pandas==2.2.2
pyarrow==16.1.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
